aerospace_ai_tool/requirements.txt
openai
google-generativeai
numpy
sympy
scipy
matplotlib
flask
python-dotenv
